# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Expense Management API",
        "description": "API documentation for managing expenses",
        "version": "1.0.0"
    },
    "paths": {
        "/api/expenses": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "List all expenses",
                "operationId": "3f05a86f8d7816bc888123589a6a443f",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Filter from date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Filter to date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with list of expenses"
                    }
                }
            },
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Create a new expense",
                "operationId": "8c3d2dfb7b738d0178bf4e278707f553",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "amount",
                                    "category",
                                    "expense_date"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Office Supplies"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 120.5
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Stationery"
                                    },
                                    "expense_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-27"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Purchased from ABC Store"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense successfully created"
                    }
                }
            }
        },
        "/api/expenses/{id}": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Get a single expense by ID",
                "operationId": "e429117f0f8dc6014b200dce65e08a7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of the expense",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense details"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Update an existing expense",
                "operationId": "f72c15420ab0c59d2fa577b40db12ce7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of the expense",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Office Supplies"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Office"
                                    },
                                    "expense_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-28"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Updated purchase details"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense successfully updated"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Delete an expense",
                "operationId": "6b6292d114c121b2130166446cc3ef13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of the expense",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Expense deleted successfully (no content)"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Expenses",
            "description": "Endpoints for managing expenses"
        }
    ]
}